requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.8
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
cycler==0.11.0
einops==0.3.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
google-auth==2.19.1
google-auth-oauthlib==0.4.6
grpcio==1.54.2
idna==3.4
imageio==2.4.1
importlib-metadata==6.6.0
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.1.2
multidict==6.0.4
numpy==1.21.6
oauthlib==3.2.2
omegaconf==2.1.0
packaging==23.1
Pillow==9.5.0
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyDeprecate==0.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.4.4
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.7.3
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.10.0+cu102
torch-fidelity==0.3.0
torchmetrics==0.5.0
torchvision==0.11.1+cu102
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==1.26.16
Werkzeug==2.2.3
yarl==1.9.2
zipp==3.15.0
